Made createComponentWithIntl a separate helper function
