+++ packages/helpers/src/index.js
@@ -5,9 +5,11 @@ import decodeHTML from "./htmlDecoder";
 import getFeed from "./getFeed";
 import getCourseFeed from "./getCourseFeed";
 import getPostFeed from "./getPostFeed";
+import createSvgIconComponent from "./createSvgIconComponent";
 
 export {
 	createComponentWithIntl,
+	createSvgIconComponent,
 	getDirectionalStyle,
 	sendRequest,
 	decodeHTML,
